Improved binary treatment
Taken a binary system with the following properties:
* both an AB and an A entry exist in SIMBAD,
* some photometry is present for the AB entry,
different behaviors of the program were expected depending on the provided input:
* AB was picked if no 'object' was given;
* AB was picked if the 'object' corresponded to the AB entry;
* A was picked if the 'object' corresponded to the A entry.

In order to avoid cross-matching problems and to improve consistency, I have introduced a function that, in such circumstances, always picks the entry corresponding to the A component. The photometry corresponding to removed AB entries is stored into a keyword named 'simbad_additional_photometry'.